requirements.txt
# Web framework dependencies
fastapi==0.108.0
uvicorn==0.25.0
openai>=1.12.0
redis==5.0.1
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.25.1

# 确保使用兼容的NumPy版本
numpy==1.26.4
chromadb>=0.4.22
tqdm>=4.65.0

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0

# 分词和文本处理
jieba==0.42.1

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Transitive dependencies (usually managed by main packages)
annotated-types==0.6.0
anyio==4.2.0
# APScheduler==3.10.4 # Removed as 'schedule' is likely used
certifi==2023.11.17
charset-normalizer==3.3.2
pydantic_core==2.14.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
h11==0.14.0
idna==3.6
click==8.1.7